+++ Packages/CrowCLI/Sources/CrowCLILib/CrowCommand.swift
@@ -79,6 +79,7 @@ public struct CrowCommand: ParsableCommand {
             HookEventCmd.self,
             CodexNotify.self,
             GenerateDocs.self,
+            Version.self,
         ]
     )